repair math module and malformed lamdas
