+++ distributions/univariate/continuous/LaplaceRand.cpp
@@ -123,7 +123,7 @@ void ShiftedAsymmetricLaplaceDistribution<RealType>::FitShift(const std::vector<
     double maxVar = *std::max_element(sample.begin(), sample.end());
     double median = this->GetSampleMean(sample);
 
-    if (!RandMath::findRootNewtonFirstOrder<double>([this, sample] (double med)
+    if (!RandMath::findRootBrentFirstOrder<double>([this, sample] (double med)
     {
         double y = 0.0;
         for (const RealType & x : sample) {
