+++ distributions/univariate/discrete/NegativeHyperGeometricRand.cpp
@@ -1,5 +1,7 @@
 #include "NegativeHyperGeometricRand.h"
 
+#include "math/GammaMath.h"
+
 template < typename IntType >
 NegativeHyperGeometricRand<IntType>::NegativeHyperGeometricRand(IntType totalSize, IntType totalSuccessesNum, IntType limitSuccessesNum)
 {
